Add http://pntest home page
