+++ src/proxy/proxy.py
@@ -1,19 +1,9 @@
-import pathlib
 import asyncio
-import os
+
 from mitmproxy import addons, master, options
 from mitmproxy.addons import termlog, keepserving, readfile
 
-def is_dev_mode():
-    return os.getenv('DEV_MODE') is not None
-
-def get_include_path():
-    if is_dev_mode():
-        app_path = pathlib.Path(__file__).parent.parent.parent.absolute()
-        return f"{app_path}/include"
-    else:
-        app_path = pathlib.Path(__file__).parent
-        return f"{app_path}/include"
+from paths import get_include_path
 
 class ErrorCheck:
     def __init__(self):
